Snaps disabled on pipe unless pipe is itself snapped
There was an issue where, if you were touching a pipe with the end of the pipe you were holding. two blue silhouettes would appear. Each pipe has a snap point at its end, so touching the end of a grabbed pipe with a pipe in the environment would trigger the snap point of both pipes. This has been fixed by making a pipe snap point inactive by default, then activating it once it has been connected to the piping system coming from the water source.

As a consequence, the only snap point enabled at the beginning of the scene is the one on the water source. When a pipe is connected to the water source, it enables the snap point on that pipe. This pipe then turns on the snap point of the next connected pipe, and this chain reaction continues indefinitely. When a pipe is removed, the snap point is disabled.