Fixed issue with teleporting through a placed pipe. Added custom input handling.
Previously, a pipe had its collider removed when it was connected to the piping system coming from the water source. This meant that if the player targeted it, it did not display as an invalid target - and further, if the player attempted to interact with it, the raycast would go through the object and cause them to teleport. Now, instead of disabling the collider, a larger non-interactable collider is added to encase the pipe until it is removed from the piping system. This fixes both issues.

Additionally, Scripts/_Input now contains an event handling system to deal with custom functionality with inputs. One must only create an event listener (ExampleListener provides a simple example of how this works) to define input functionality. A more detailed description and exact operation can be seen at https://www.youtube.com/watch?v=NdHGuj-u-kc