Fixed all elbows rotating, reconfigured scripts.
All elbows no longer rotate when one is selected to rotate. For some reason, when elbow is first connected, it rotates 90 degrees per button press - but once you connect something to it, this elbow (and all subsequent elbows) rotate 180 degrees.

Functionality scripting was messy and confusing and has been refactored to a Pipe class, as well as inherited classes for the different kind of pipes. These are attached to SnapPoint child GameObject of a pipe - in the future, it should probably be moved to the root GameObject.